requirements.txt
streamlit
openai
streamlit-js-eval
streamlit-folium
folium
pandas
fpdf
python-dotenv
openpyxl